some more refactoring. Extracted a class UserInterfaceModel and gave it all methods related to the UI
