+++ src/mcmfuserinterface/drag_drop_table/columns/AssignedProjectsCountColumn.java
@@ -7,12 +7,12 @@ package mcmfuserinterface.drag_drop_table.columns;
 
 import java.util.Comparator;
 
+import ford_fulkerson.model.Reader;
 import javafx.beans.property.ReadOnlyObjectWrapper;
 import javafx.scene.control.Label;
 import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
 import mcmfuserinterface.drag_drop_table.TableObjectInterface;
-import model.Reader;
 
 /**
  *
