+++ src/mcmfuserinterface/drag_drop_table/columns/ListColumn.java
@@ -5,11 +5,17 @@
  */
 package mcmfuserinterface.drag_drop_table.columns;
 
+import ford_fulkerson.model.Project;
+import ford_fulkerson.model.Reader;
 import javafx.beans.property.ReadOnlyObjectWrapper;
+import javafx.scene.control.Label;
+import javafx.scene.control.ScrollPane;
+import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
+import javafx.scene.layout.HBox;
 import mcmfuserinterface.controllers.ControllerInterface;
-import mcmfuserinterface.drag_drop_table.TableCellWithListFactory;
 import mcmfuserinterface.drag_drop_table.TableObjectInterface;
+import mcmfuserinterface.drag_drop_table.components.DroppableScrollPane;
 
 /**
  *
@@ -18,7 +24,7 @@ import mcmfuserinterface.drag_drop_table.TableObjectInterface;
 public class ListColumn  extends TableColumn<TableObjectInterface, TableObjectInterface>{
     
     @SuppressWarnings({ "rawtypes", "unchecked" })
-	public ListColumn(String name, ControllerInterface contr){
+	public ListColumn(String name, ControllerInterface controller){
         super(name);
         setMinWidth(350);
         setPrefWidth(450);
@@ -28,7 +34,33 @@ public class ListColumn  extends TableColumn<TableObjectInterface, TableObjectIn
               return new ReadOnlyObjectWrapper(features.getValue());
         });
         
-        setCellFactory(new TableCellWithListFactory(contr));
+        setCellFactory( item ->{
+        	return new TableCell<TableObjectInterface, TableObjectInterface>() {
+                @Override
+                public void updateItem(final TableObjectInterface object, boolean empty) {
+                    if(!empty && object != null){
+                    	
+                    	HBox hbox = new HBox();
+                        hbox.setSpacing(10);
+                        hbox.setUserData(object);
+                        
+                        Reader reader = (Reader) object;
+                        for (Project project : controller.getReaderList(reader)) {
+                            Label label = controller.createLabel(reader, project);
+                            hbox.getChildren().add(label);
+                        }
+                        
+                        ScrollPane scrollPane = new DroppableScrollPane(controller);
+                        scrollPane.setContent(hbox);
+                        scrollPane.setContextMenu(controller.createContextMenu((Reader)object,hbox));
+                        
+                        this.setGraphic(scrollPane);
+                    } else {
+                        setGraphic(null);
+                    }
+                }
+            };
+        });
     }
     
 }
