+++ src/test/Constraints_Test.java
@@ -6,16 +6,15 @@ import static org.junit.Assert.assertTrue;
 
 import java.util.ArrayList;
 
-import model.MCMFModel;
-import model.Project;
-import model.Reader;
-
 import org.junit.Test;
 
 import test.graph_creator.RandomArbitraryModel;
 import test.graph_creator.RandomReaderAllocationModel;
 import ford_fulkerson.MinCostMaxFlowAlgorithm;
 import ford_fulkerson.ReaderShortlistException;
+import ford_fulkerson.model.MCMFModel;
+import ford_fulkerson.model.Project;
+import ford_fulkerson.model.Reader;
 import ford_fulkerson.network.Edge;
 import ford_fulkerson.network.Network;
